Rename 'Health API' to 'Actuator API'
