+++ docs/Interfaces-Data-Sources.md
@@ -21,7 +21,7 @@ The **OTP Transit Index API** is another [RESTful](https://en.wikipedia.org/wiki
 
 Additional experimental APIs are provided by [sandbox extensions](SandboxExtension.md):
 
-The [Health API](sandbox/ActuatorAPI.md) provides endpoints for checking the health status of the OTP instance. It can be useful when running OTP in a container.
+The [Actuator API](sandbox/ActuatorAPI.md) provides endpoints for checking the health status of the OTP instance. It can be useful when running OTP in a container.
 
 The [Transmodel GraphQL API](sandbox/TransmodelApi.md) is the Transmodel API (version 3) used at Entur in production(Sep, 2020).
