+++ docs/SandboxExtension.md
@@ -11,7 +11,7 @@ provided "as is".
 
 - [Examples](sandbox/Examples.md) - Sandbox examples on how to implement extensions.
 - [Google Cloud Storage](sandbox/GoogleCloudStorage.md) - Enable Google Cloud Storage as a OTP Data Source
-- [Health API](sandbox/ActuatorAPI.md) - API used to check the health status of the OTP instance.
+- [Actuator API](sandbox/ActuatorAPI.md) - API used to check the health status of the OTP instance.
 - [Transfer analyser](sandbox/transferanalyzer.md) - Module used for analyzing the transfers between 
 nearby stops generated by routing via OSM data.
 - [HSL Legacy GraphQL API](sandbox/LegacyGraphQLApi.md) - HSL's GraphQL API used by the Digitransit project.
